requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

#-- Database ---

#sing stable v3 client to avoid v4 init issues

weaviate-client==3.26.7

#-- Ingestion ---

llama-parse>=0.4.0
nest_asyncio>=1.6.0

#-- AI / ML ---

google-generativeai>=0.3.0  # Embeddings
groq>=0.4.0                 # Inference (Future)
numpy>=1.26